more developed start menu

+++ src/taskbar/app/StartButton.tsx
@@ -1,21 +1,23 @@
 import type {ProgramMainData} from "ollieos/types";
 import type {SpawnResult} from "ollieos/kernel";
 
-import {useCallback, useRef} from "react";
+import {useCallback, useMemo, useRef} from "react";
 import {LayoutPanelLeft} from "lucide-react";
 
 export const StartButton = ({main_data}: {main_data: ProgramMainData}) => {
-    const {kernel} = main_data;
+    const {kernel, shell} = main_data;
 
+    const proc_mgr = useMemo(() => kernel.get_process_manager(), [kernel]);
     const start_menu_process = useRef<SpawnResult | null>(null);
 
     const on_click = useCallback(
         () => {
-            if (start_menu_process.current !== null) {
+            // TODO: once dead flag added to process, remove this bypass
+            if (start_menu_process.current !== null && proc_mgr.list_pids().includes(start_menu_process.current.process.pid)) {
                 start_menu_process.current.process.kill(0);
                 start_menu_process.current = null;
             } else {
-                start_menu_process.current = kernel.spawn("sl_startmenu");
+                start_menu_process.current = kernel.spawn("sl_startmenu", [], shell);
             }
         },
         [kernel]
